# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 班级网站网站 API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /knowledge/imports:
    get:
      summary: 获取当前成员可管理的知识导入任务
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses:
        "200":
          description: 导入任务分页
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnowledgeImportJobPage" }
    post:
      summary: 上传 Markdown 或最大 200 MiB 的 ZIP 知识包
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, libraryId]
              properties:
                file: { type: string, format: binary }
                libraryId: { type: string }
                targetDocumentId: { type: string }
      responses:
        "202":
          description: 已进入异步预检队列
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnowledgeImportJob" }
        "413": { description: 文件超过限制 }
  /knowledge/imports/{id}:
    get:
      summary: 获取知识导入任务
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: 导入任务
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnowledgeImportJob" }
    delete:
      summary: 取消知识导入任务
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: 已取消或已请求 Worker 取消 }
  /knowledge/imports/{id}/structure:
    get:
      summary: 分页获取预检后的知识树摘要
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - in: query
          name: cursor
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
      responses:
        "200":
          description: 节点摘要和图片引用
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnowledgeImportStructure" }
  /knowledge/imports/{id}/issues:
    get:
      summary: 分页获取知识包预检问题
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses:
        "200":
          description: 预检问题分页
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnowledgeImportIssuePage" }
  /knowledge/imports/{id}/confirm:
    post:
      summary: 确认预检结果、原子预留容量并进入索引队列
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: 已确认
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnowledgeImportJob" }
  /knowledge/imports/{id}/retry-compensation:
    post:
      summary: 显式重试补偿失败任务的精确对象清理
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: 已重新进入补偿清理队列
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnowledgeImportJob" }
        "400": { $ref: "#/components/responses/ValidationError" }
...
  /knowledge/{id}:
    get:
      summary: 获取已发布资料详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 资料详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnowledgeDocumentDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: 删除知识资料、检索分片和原始上传对象
      description: 数据库先执行软删除；对象存储清理失败时保留对象键并返回 storageDeleted=false，允许后续重试。
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 资料已从知识库删除
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KnowledgeDeleteResult" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /quizzes/imports/{id}:
    get:
      summary: 获取导入任务、预检摘要与前 100 条问题
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 导入任务详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizImportJobDetail" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /quizzes/imports/{id}/confirm:
    post:
      summary: 幂等确认预检通过的导入任务
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201":
          description: 已确认或返回既有处理结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizImportJobDetail" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /quizzes/questions/{id}:
    get:
      summary: 获取包含答案与评分标准的题目编辑数据
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 题目编辑数据
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizQuestionEditorData" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: 更新题目及其多章节关联
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuizQuestionWriteRequest" }
      responses:
        "200":
          description: 题目已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizQuestionEditorData" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      summary: 停用题目并保留历史答题快照
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 题目已停用；重复调用保持幂等成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizQuestionDeleteResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /ai/question-reviews:
    get:
      summary: 分页列出 AI 题目审核队列
      x-required-roles: [EDITOR, ADMIN]
      responses:
        "200":
          description: 审核队列
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiQuestionReviewPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /ai/question-reviews/{questionId}:
    get:
      summary: 获取 AI 题当前稿、原始生成稿、精确证据和审核事件
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: questionId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 审核详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiQuestionReviewDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: 按审核版本保存 AI 题草稿
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: questionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AiQuestionReviewWriteRequest" }
      responses:
        "200": { description: 已保存草稿 }
        "409": { $ref: "#/components/responses/Conflict" }
  /ai/question-reviews/{questionId}/approve:
    post:
      summary: 原子保存并发布单道 AI 题
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: questionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AiQuestionReviewWriteRequest" }
      responses:
        "201": { description: 已发布题目 }
        "409": { $ref: "#/components/responses/Conflict" }
  /ai/question-reviews/{questionId}/reject:
    post:
      summary: 按审核版本拒绝单道 AI 题
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: questionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AiQuestionRejectRequest" }
      responses:
        "201": { description: 已拒绝题目 }
        "409": { $ref: "#/components/responses/Conflict" }
  /ai/question-reviews/{questionId}/reopen:
    post:
      summary: 按审核 revision 将已批准 AI 题显式撤回至待审核草稿
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: questionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AiQuestionReopenRequest" }
      responses:
        "201":
          description: 已撤回；新练习与每日计划不再选择该题
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiQuestionReviewDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /ai/question-reviews/revalidate-source/bulk:
    post:
      summary: 批量自动重绑待复核 AI 题的当前知识来源
      x-required-roles: [EDITOR, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AiQuestionSourceBulkRevalidateRequest" }
      responses:
        "201":
          description: 逐题重绑结果；单题失败不影响其他题目
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiQuestionSourceBulkRevalidateResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
  /ai/question-reviews/{questionId}/revalidate-source:
    post:
      summary: 将 AI 题重新绑定到当前有效的显式知识节点
      x-required-roles: [EDITOR, ADMIN]
      parameters:
        - in: path
          name: questionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AiQuestionSourceRevalidateRequest" }
      responses:
        "201": { description: 来源已复核 }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bmc3_session
  parameters:
...
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
...
  responses:
...
    Unauthorized:
      description: 未登录、会话失效或凭据无效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: 当前账号无权限或操作被安全规则拒绝
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: 目标不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Conflict:
      description: 操作与当前资源状态冲突
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    ValidationError:
      description: 请求体校验失败
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    KnowledgeImportStatus:
      type: string
      enum: [PREFLIGHT_PENDING, PREFLIGHTING, INVALID, AWAITING_CONFIRMATION, INDEX_PENDING, PROCESSING, READY, FAILED, COMPENSATION_FAILED, EXPIRED]
    ImportSourceCleanupStatus:
      type: string
      enum: [RETAINED, PENDING, IN_PROGRESS, CLEANED, FAILED]
...
    KnowledgeImportJob:
      type: object
      required: [id, libraryId, fileType, sourceName, sourceSize, status, stage, nodeCount, estimatedChunkCount, imageCount, warningCount, errorCount, sourceCleanupStatus, sourceCleanupAttempts, sourceCleanupNextAttemptAt, sourceCleanupError, sourceCleanedAt, cancelRequestedAt, expiresAt, createdAt, updatedAt]
      properties:
        id: { type: string }
        libraryId: { type: string }
        targetDocumentId: { type: [string, "null"] }
        fileType: { type: string, enum: [MARKDOWN, ZIP] }
        sourceName: { type: string }
        sourceSize: { type: integer, maximum: 209715200 }
        status: { $ref: "#/components/schemas/KnowledgeImportStatus" }
        stage: { type: string }
        progressCurrent: { type: integer, minimum: 0 }
        progressTotal: { type: integer, minimum: 0 }
        title: { type: [string, "null"] }
        nodeCount: { type: integer, minimum: 0 }
        estimatedChunkCount: { type: integer, minimum: 0 }
        imageCount: { type: integer, minimum: 0 }
        processedImageBytes: { type: integer, minimum: 0 }
        warningCount: { type: integer, minimum: 0 }
        errorCount: { type: integer, minimum: 0 }
        errorCode: { type: [string, "null"] }
        errorMessage: { type: [string, "null"] }
        sourceCleanupStatus: { $ref: "#/components/schemas/ImportSourceCleanupStatus" }
        sourceCleanupAttempts: { type: integer, minimum: 0, maximum: 5 }
        sourceCleanupNextAttemptAt: { type: [string, "null"], format: date-time }
        sourceCleanupError: { type: [string, "null"] }
        sourceCleanedAt: { type: [string, "null"], format: date-time }
        cancelRequestedAt: { type: [string, "null"], format: date-time }
        confirmedVersionId: { type: [string, "null"] }
        expiresAt: { type: string, format: date-time }
        confirmedAt: { type: [string, "null"], format: date-time }
        completedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    KnowledgeImportJobPage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/KnowledgeImportJob" } }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
    KnowledgeImportIssue:
      type: object
      required: [id, severity, code, message, entryPath, nodePath, line, column, createdAt]
      properties:
        id: { type: integer, minimum: 1 }
        severity: { type: string, enum: [ERROR, WARNING] }
        code: { type: string, maxLength: 80 }
        message: { type: string, maxLength: 500 }
        entryPath: { type: [string, "null"], maxLength: 500 }
        nodePath: { type: [string, "null"], maxLength: 500 }
        line: { type: [integer, "null"], minimum: 1 }
        column: { type: [integer, "null"], minimum: 1 }
        createdAt: { type: string, format: date-time }
    KnowledgeImportIssuePage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/KnowledgeImportIssue" } }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
    KnowledgeImportStructure:
      type: object
      required: [items, nextCursor, total, images, titleMarkdown, replacement]
      properties:
        items:
          type: array
          items:
            type: object
            required: [level, title, titleMarkdown, path, chapterName, chunkCount]
            properties:
              level: { type: integer, minimum: 2, maximum: 6 }
              title: { type: string }
              titleMarkdown: { type: string, maxLength: 4096 }
              path: { type: string }
              chapterName: { type: string }
              chunkCount: { type: integer, minimum: 0 }
        nextCursor: { type: [integer, "null"], minimum: 0 }
        total: { type: integer, minimum: 0 }
        images:
          type: array
          items:
            type: object
            required: [entryPath, altText, nodePathHash, occurrenceIndex]
            properties:
              entryPath: { type: string }
              altText: { type: string, minLength: 1, maxLength: 300 }
              nodePathHash: { type: string, minLength: 64, maxLength: 64 }
              occurrenceIndex: { type: integer, minimum: 0 }
        titleMarkdown: { type: [string, "null"], maxLength: 4096 }
        replacement:
          oneOf:
            - { type: "null" }
            - { $ref: "#/components/schemas/KnowledgeImportReplacementPreview" }
    KnowledgeImportReplacementPreview:
      type: object
      additionalProperties: false
      required: [mode, targetDocumentId, targetDocumentTitle, targetDocumentCreatedAt, activeVersionId, matchedH2Titles]
      properties:
        mode: { type: string, enum: [CREATE, REPLACE] }
        targetDocumentId: { type: [string, "null"] }
        targetDocumentTitle: { type: [string, "null"] }
        targetDocumentCreatedAt: { type: [string, "null"], format: date-time }
        activeVersionId: { type: [string, "null"] }
        matchedH2Titles:
          type: array
          items: { type: string, maxLength: 200 }
...
    KnowledgeDocument:
      type: object
      required:
        [
          id,
          title,
          kind,
          indexStatus,
          sourceName,
          mimeType,
          fileSize,
          subject,
          publishedAt,
        ]
      properties:
        id: { type: string }
        title: { type: string, maxLength: 200 }
        kind: { $ref: "#/components/schemas/KnowledgeKind" }
        indexStatus: { $ref: "#/components/schemas/IndexStatus" }
        sourceName: { type: [string, "null"] }
        mimeType: { type: [string, "null"] }
        fileSize: { type: integer, minimum: 0, maximum: 209715200 }
        subject: { $ref: "#/components/schemas/KnowledgeSubject" }
        publishedAt: { type: [string, "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
    KnowledgeDocumentDetail:
      allOf:
        - { $ref: "#/components/schemas/KnowledgeDocument" }
        - type: object
          required: [body]
          properties:
            body: { type: [string, "null"] }
...
    KnowledgeDeleteResult:
      type: object
      required: [id, deleted, storageDeleted]
      properties:
        id: { type: string }
        deleted: { type: boolean, const: true }
        storageDeleted: { type: boolean }
    ApiError:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        code: { type: string }
        message:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
        requestId: { type: string }
...
    QuestionType:
      type: string
      enum: [SINGLE, MULTIPLE, TRUE_FALSE, SHORT_ANSWER]
...
    QuizPastPaper:
      type: object
      required: [id, title, subjectId, subject, year]
      properties:
        id: { type: string }
        title: { type: string }
        subjectId: { type: string }
        subject: { type: string }
        year: { type: [integer, "null"] }
...
    QuizQuestionWriteRequest:
      type: object
      additionalProperties: false
      required:
        [
          gradingType,
          subjectId,
          chapterIds,
          prompt,
          options,
          correctAnswer,
          explanation,
        ]
      properties:
        gradingType: { $ref: "#/components/schemas/QuestionType" }
        typeLabel: { type: string, minLength: 1, maxLength: 100 }
        subjectId: { type: string }
        chapterIds:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items: { type: string }
        category: { $ref: "#/components/schemas/QuizQuestionCategory" }
        prompt: { type: string, minLength: 2, maxLength: 10000 }
        options:
          type: array
          items: { $ref: "#/components/schemas/QuizOption" }
        correctAnswer: { type: array, items: { type: string } }
        gradingRubric: { $ref: "#/components/schemas/GradingRubric" }
        explanation: { type: string, maxLength: 10000 }
        pastPaperId: { type: string }
        paperOrder: { type: integer, minimum: 1, maximum: 10000 }
    QuizQuestionEditorData:
      allOf:
        - { $ref: "#/components/schemas/QuizQuestionSummary" }
        - type: object
          required: [correctAnswer, explanation]
          properties:
            correctAnswer: { type: array, items: { type: string } }
            gradingRubric: { $ref: "#/components/schemas/GradingRubric" }
            explanation: { type: string }
    QuizQuestionDeleteResult:
      type: object
      required: [id, deleted]
      properties:
        id: { type: string }
        deleted: { type: boolean, const: true }
...
    QuizImportIssue:
      type: object
      required: [id, severity, code, rowNumber, filePath, field, message, createdAt]
      properties:
        id: { type: integer }
        severity: { type: string, enum: [ERROR, WARNING] }
        code: { type: string }
        rowNumber: { oneOf: [{ type: integer }, { type: "null" }] }
        filePath: { oneOf: [{ type: string }, { type: "null" }] }
        field: { oneOf: [{ type: string }, { type: "null" }] }
        message: { type: string }
        createdAt: { type: string, format: date-time }
...
    QuizImportJobDetail:
      allOf:
        - { $ref: "#/components/schemas/QuizImportJob" }
        - type: object
          properties:
            issues:
              type: array
              items: { $ref: "#/components/schemas/QuizImportIssue" }
            issuesTruncated: { type: boolean }
            resultPastPaper:
              oneOf:
                - { $ref: "#/components/schemas/QuizPastPaper" }
                - { type: "null" }
...
    AiQuestionReviewPage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AiQuestionReviewSummary" } }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
    AiQuestionReviewDetail:
      allOf:
        - { $ref: "#/components/schemas/AiQuestionReviewSummary" }
        - type: object
          required: [gradingType, options, correctAnswer, maxScore, explanation, sourceRevision, originalGenerated, sources, events]
          properties:
            gradingType: { $ref: "#/components/schemas/QuestionType" }
            options: { type: array, items: { $ref: "#/components/schemas/QuizOption" } }
            correctAnswer: { type: array, items: { type: string } }
            gradingRubric: { oneOf: [{ type: object, additionalProperties: true }, { type: "null" }] }
            maxScore: { type: integer, minimum: 1 }
            explanation: { type: string }
            sourceRevision: { type: integer, minimum: 1 }
            originalGenerated: {}
            sources:
              type: array
              items:
                type: object
                required: [id, sourceRevision, ordinal, current, libraryId, documentId, documentVersionId, libraryChapterId, title, breadcrumb, contentHash, evidenceContent]
                properties:
                  id: { type: string }
                  sourceRevision: { type: integer, minimum: 1 }
                  ordinal: { type: integer, minimum: 0 }
                  current: { type: boolean }
                  knowledgeNodeId: { oneOf: [{ type: string }, { type: "null" }] }
                  libraryId: { type: string }
                  documentId: { type: string }
                  documentVersionId: { type: string }
                  libraryChapterId: { type: string }
                  title: { type: string }
                  breadcrumb: { type: string }
                  contentHash: { type: string }
                  evidenceContent: { type: string }
            events: { type: array, items: { type: object, additionalProperties: true } }
    AiQuestionReviewWriteRequest:
      type: object
      additionalProperties: false
      required: [expectedRevision, typeLabel, chapterIds, prompt, options, correctAnswer, explanation]
      properties:
        expectedRevision: { type: integer, minimum: 1 }
        typeLabel: { type: string, minLength: 1, maxLength: 100 }
        chapterIds: { type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: { type: string } }
        prompt: { type: string, minLength: 2, maxLength: 10000 }
        options: { type: array, maxItems: 20, items: { $ref: "#/components/schemas/QuizOption" } }
        correctAnswer: { type: array, minItems: 1, maxItems: 20, items: { type: string } }
        gradingRubric: { oneOf: [{ type: object, additionalProperties: true }, { type: "null" }] }
        explanation: { type: string, maxLength: 10000 }
    AiQuestionRejectRequest:
      type: object
      additionalProperties: false
      required: [expectedRevision, reason]
      properties:
        expectedRevision: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 1, maxLength: 1000 }
    AiQuestionReopenRequest:
      type: object
      additionalProperties: false
      required: [expectedRevision]
      properties:
        expectedRevision: { type: integer, minimum: 1 }
    AiQuestionSourceRevalidateRequest:
      type: object
      additionalProperties: false
      required: [expectedRevision, knowledgeNodeIds]
      properties:
        expectedRevision: { type: integer, minimum: 1 }
        knowledgeNodeIds: { type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: { type: string } }
    AiQuestionSourceBulkRevalidateRequest:
      type: object
      additionalProperties: false
      required: [questionIds]
      properties:
        questionIds: { type: array, minItems: 1, maxItems: 50, items: { type: string, minLength: 1, maxLength: 191 } }
...
    AiQuestionSourceBulkRevalidateResponse:
      type: object
      additionalProperties: false
      required: [results]
      properties:
        results:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/AiQuestionSourceBulkRevalidateResult" }